Changes to run tests on Ubuntu 15.04 with 3.19 Kernel - Started prettifying code and make it more clear

+++ vm.py
@@ -2,11 +2,27 @@
 # File I/O
 import sys
 import os
+import logging
 import time
 import subprocess as sub
 import shutil
 import tarfile
 from subprocess import check_output
+
+###############
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+# create handler for filelogging
+handler = logging.FileHandler('vm.log')
+handler.setLevel(logging.INFO)
+
+# create a logging format
+formatter = logging.Formatter('%(levelname)s - %(name)s - %(message)s')
+handler.setFormatter(formatter)
+
+logger.addHandler(handler)
+###############
 
 workingDir = "/mnt/"
 filesPath  = "/mnt/files"
@@ -22,10 +38,10 @@ def evertestExtractTest(dottest, testname):
 		if tarfile.is_tarfile(extractString):
 			tfile.extractall(extractPath)
 		else:
-			print extractString + " is not a tarfile."
+			logger.error("%s is not a tarfile.", extractString)
 	except:
 		e = sys.exc_info()[1]
-		print "Error in evertestExtractTest: \n" + str(e)
+		logger.error("Error in evertestExtractTest: \n %s", str(e))
 
 getHostname = "hostname"
 hostname = check_output(getHostname)
@@ -43,24 +59,35 @@ scriptFile = hostname + ".script"
 sourceDir = "/mnt/" + testname + "/scripts/"
 destination = workingDir
 source = sourceDir + scriptFile
-shutil.copy(source, destination)
+if (os.path.exists(source)):
+	shutil.copy(source, destination)
+else:
+	logger.error("%s cannot be found!", source)
 
 #Config File
 configDir = "/mnt/" + testname + "/"
 config = configDir + testname + ".conf"
 source = config
 destination = workingDir
-shutil.copy(source, destination)
+if (os.path.exists(source)):
+	shutil.copy(source, destination)
+else:
+	logger.error("%s cannot be found!", source)
 
 #Other files
 srcFilesDir = "/mnt/" + testname + "/files/"
 if os.path.exists(srcFilesDir):
 	shutil.copytree(srcFilesDir, filesPath, symlinks=False, ignore=None)
+else:
+	logger.error("%s does not exist!", srcFilesDir)
 
 #Execute script
 scriptFile = "/mnt/" + hostname + ".script"
 
 runs = "python -B " + scriptFile
-run = sub.Popen(runs, shell=True, stdout=sub.PIPE)
+try:
+	run = sub.Popen(runs, shell=True, stdout=sub.PIPE)
+except:
+	logger.error("Error on executing '%s'!", runs)
 run.wait()
